% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description_example_simulation.R
\docType{data}
\name{example_simulation}
\alias{example_simulation}
\title{Example Simulation Results}
\format{
A tibble with 10 rows and 5 variables:
\describe{
\item{split}{Repetition number of the simulation}
\item{sim_n}{Simulation trial number}
\item{DURATION}{Duration year of the policy}
\item{PREM}{Premium}
\item{CLAIM}{Claim}
}
}
\usage{
data(example_simulation)
}
\description{
A synthetic dataset representing simulated insurance cash flows over time.
This dataset is provided for demonstration and validation of key package functions,
including risk evaluation (\code{risk()}) and profit commission calculation (\code{calc_pc()}).
It includes premium and claim flows per simulation trial and policy duration,
enabling users to test volatility risk metrics and profit commission calculations.
}
\keyword{datasets}
